refactor(chat): split chat.go into focused files
